requirements.txt
alembic==1.7.5
aniso8601==9.0.1
autopep8==1.6.0
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpickle==2.0.0
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typing_extensions==4.0.0
Werkzeug==2.0.2
